test: checkpoint kurtosis devnet deployment + test cases
