--- world-chain-builder/crates/assertor/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "assertor"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tracing = "0.1"
-tracing-subscriber = {version = "0.3.18", features = ["env-filter"]}
-clap = { version = "4", features = ["derive", "env"] }
-alloy-primitives = "0.8"
-alloy-provider = "0.4"
-alloy-transport = "0.4"
-alloy-network = "0.4"
-alloy-rpc-types-eth = "0.4"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-futures = "0.3"
-tokio = { version = "1", features = ["full"] }
-eyre = { version = "0.6", package = "color-eyre" }
