+++ world-chain-builder/crates/tests/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "tests"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+# world
+world-chain-builder-pool = { workspace = true, features = ["test"] }
+world-chain-builder-node = {workspace = true, features = ["test"] }
+
+# alloy
+alloy-primitives.workspace = true
+alloy-provider.workspace = true
+alloy-transport.workspace = true
+alloy-network.workspace = true
+alloy-rpc-types-eth.workspace = true
+
+# async
+tokio = { workspace = true, features = ["full"] }
+futures.workspace = true
+
+# third-party
+serde_json.workspace = true
+serde.workspace = true
+eyre.workspace = true
+tracing-subscriber = { workspace = true, features = ["env-filter"]}
+tracing.workspace = true
+
+[lints]
+workspace = true
