+++ world-chain-builder/crates/world/node/Cargo.toml
@@ -13,6 +13,8 @@ workspace = true
 [dependencies]
 world-chain-builder-payload.workspace = true
 world-chain-builder-pool.workspace = true
+world-chain-builder-rpc = { workspace = true, optional = true }
+world-chain-builder-pbh = { workspace = true, optional = true }
 
 reth.workspace = true
 reth-chain-state.workspace = true
@@ -34,10 +36,23 @@ alloy-primitives.workspace = true
 alloy-rpc-types.workspace = true
 alloy-sol-types.workspace = true
 
+reth-e2e-test-utils = { workspace = true, optional = true }
+reth-evm = { workspace = true, optional = true }
+reth-node-core = { workspace = true, optional = true }
+reth-optimism-consensus = { workspace = true, optional = true }
+reth-optimism-evm = { workspace = true, optional = true }
+revm-primitives = { workspace = true, optional = true }
+
+alloy-genesis = { workspace = true, optional = true }
+alloy-network = { workspace = true, optional = true }
+alloy-signer-local = { workspace = true, optional = true }
+
 tokio.workspace = true
 eyre.workspace = true
 futures.workspace = true
 clap.workspace = true
+serde_json = { workspace = true, optional = true}
+chrono = { workspace = true, optional = true }
 
 [dev-dependencies]
 world-chain-builder-pbh.workspace = true
@@ -62,3 +77,20 @@ semaphore.workspace = true
 serde_json.workspace = true
 chrono.workspace = true
 
+[features]
+default = []
+test = [
+    "dep:world-chain-builder-pbh",
+    "dep:world-chain-builder-rpc",
+    "dep:reth-e2e-test-utils",
+    "dep:reth-evm",
+    "dep:reth-node-core",
+    "dep:reth-optimism-consensus",
+    "dep:reth-optimism-evm",
+    "dep:revm-primitives",
+    "dep:alloy-genesis",
+    "dep:alloy-network",
+    "dep:alloy-signer-local",
+    "dep:chrono",
+    "dep:serde_json",
+]
