+++ world-chain-builder/crates/world/node/src/tests/assets/genesis.json
