+++ world-chain-builder/crates/world/node/src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod e2e;
+pub mod test_utils;
